+++ doc/changelog.md
@@ -14,6 +14,7 @@
 - Added `std/wbmp`.
 - Added `WUFFS_BASE__PIXEL_BLEND__SRC_OVER`.
 - Added `WUFFS_BASE__PIXEL_FORMAT__BGR_565`.
+- Added double-curly blocks.
 - Added interfaces.
 - Added preprocessor.
 - Added single-quoted strings.